+++ spec/lib/v1_0/patient_statement_spec.rb
@@ -1,27 +1,28 @@
 describe 'Eligible::PatientStatement' do
-  let(:params) { { test: true, rest_api_version: '1.0' } }
+  let(:params) { { rest_api_version: '1.0' } }
   let(:api_key) { 'xyz' }
   let(:response) { { success: true } }
   before(:each) do
     allow(Eligible::Util).to receive(:convert_to_eligible_object).with(response, api_key).and_return('success')
+    allow(Eligible::PatientStatement).to receive(:endpoint_name).and_return('patient_statements')
   end
 
-  describe '.get' do
+  describe '.retrieve' do
     it 'should call Eligible.request with proper url' do
-      params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+      params[:id] = 'pst_9bcb7c733e0242439575a299'
       allow(Eligible).to receive(:request).with(:get, '/patient_statements/pst_9bcb7c733e0242439575a299', api_key, params, {}).and_return([response, api_key])
-      expect(Eligible::PatientStatement.get(params, api_key: api_key)).to eq 'success'
+      expect(Eligible::PatientStatement.retrieve(params[:id], api_key: api_key)).to eq 'success'
     end
 
     it 'should raise error if patient_statement id is not present' do
-      expect { Eligible::PatientStatement.get(params, api_key: api_key) }.to raise_error(ArgumentError)
+      expect { Eligible::PatientStatement.retrieve(nil, api_key: api_key) }.to raise_error(ArgumentError)
     end
   end
 
-  describe '.post' do
+  describe '.create' do
     it 'should post to Eligible.request with proper url' do
       allow(Eligible).to receive(:request).with(:post, '/patient_statements', api_key, params, {}).and_return([response, api_key])
-      expect(Eligible::PatientStatement.post(params, api_key: api_key)).to eq 'success'
+      expect(Eligible::PatientStatement.create(params, api_key: api_key)).to eq 'success'
     end
   end
 
@@ -34,7 +35,7 @@ describe 'Eligible::PatientStatement' do
 
   describe '.update' do
     it 'should call Eligible.request with proper url' do
-      params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+      params[:id] = 'pst_9bcb7c733e0242439575a299'
       allow(Eligible).to receive(:request).with(:put, '/patient_statements/pst_9bcb7c733e0242439575a299', api_key, params, {}).and_return([response, api_key])
       expect(Eligible::PatientStatement.update(params, api_key: api_key)).to eq 'success'
     end
@@ -52,19 +53,21 @@ describe 'Eligible::PatientStatement' do
 
     context '.finalize' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/finalize', api_key, params, {}).and_return([response, api_key])
-        expect(Eligible::PatientStatement.finalize(params, api_key: api_key)).to eq 'success'
+        expect(Eligible::PatientStatement.finalize(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        expect { Eligible::PatientStatement.finalize(params, api_key: api_key) }.to raise_error(ArgumentError)
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        expect { Eligible::PatientStatement.finalize(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
 
     context '.payment' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/payment', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::PatientStatement.payment(params, api_key: api_key)).to eq 'success'
       end
@@ -76,7 +79,7 @@ describe 'Eligible::PatientStatement' do
 
     context '.payments' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
         allow(Eligible).to receive(:request).with(:get, '/patient_statements/pst_9bcb7c733e0242439575a299/payments', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::PatientStatement.payments(params, api_key: api_key)).to eq 'success'
       end
@@ -88,7 +91,7 @@ describe 'Eligible::PatientStatement' do
 
     context '.process' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/process', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::PatientStatement.process(params, api_key: api_key)).to eq 'success'
       end
@@ -100,7 +103,7 @@ describe 'Eligible::PatientStatement' do
 
     context '.capture' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/capture', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::PatientStatement.capture(params, api_key: api_key)).to eq 'success'
       end
@@ -112,24 +115,56 @@ describe 'Eligible::PatientStatement' do
 
     context '.send' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/send', api_key, params, {}).and_return([response, api_key])
-        expect(Eligible::PatientStatement.send(params, api_key: api_key)).to eq 'success'
+        expect(Eligible::PatientStatement.send(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        expect { Eligible::PatientStatement.send(params, api_key: api_key) }.to raise_error(ArgumentError)
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        expect { Eligible::PatientStatement.send(nil, api_key: api_key) }.to raise_error(ArgumentError)
+      end
+    end
+
+    context '.void' do
+      it 'should call Eligible.request with proper url' do
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/void', api_key, params, {}).and_return([response, api_key])
+        expect(Eligible::PatientStatement.void(params[:id], api_key: api_key)).to eq 'success'
+      end
+
+      it 'should raise error if patient_statement id is not present' do
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        expect { Eligible::PatientStatement.void(nil, api_key: api_key) }.to raise_error(ArgumentError)
+      end
+    end
+
+    context '.mark_uncollectible' do
+      it 'should call Eligible.request with proper url' do
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/mark_uncollectible', api_key, params, {}).and_return([response, api_key])
+        expect(Eligible::PatientStatement.mark_uncollectible(params[:id], api_key: api_key)).to eq 'success'
+      end
+
+      it 'should raise error if patient_statement id is not present' do
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
+        expect { Eligible::PatientStatement.mark_uncollectible(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
 
     context '.reestimate' do
       it 'should call Eligible.request with proper url' do
-        params[:patient_statement_id] = 'pst_9bcb7c733e0242439575a299'
+        params[:id] = 'pst_9bcb7c733e0242439575a299'
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/reestimate', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::PatientStatement.reestimate(params, api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
+        allow(Eligible::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::PatientStatement.reestimate(params, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end